Updated import renamed function to underscores added better error handling
